Available files in repertoire view dropdown list are ordered alphabetically
When more than just a few files are in the share folder, attaching files to a song in the repertoire becomes very tedious. The user has to scroll and read most of the list, because the files are not ordered. This fix improves the usability a lot by sorting the files alphabetically.

Signed-off-by: Dennis Noppeney <info@dennis-noppeney.de>