Fixed growing zip archives
When a zip archive is created from a shared folder shortly after another zip archive has been created, the new archive contains the selected files and also the files selected for the previous zip archive. This behaviour makes zip archives grow larger and larger with every request.

The reason seems to be the ZipArchive::CREATE option, which only creates a new archive if  it does not exist. If the archive exists, the exist it is not deleted. The option ZipArchive::OVERWRITE seems more suitable.