Fixed zip download of Share Folders
The fix cf8107e3 from November 2019 fixed the unexpected big zip archives due to unintended appending of new content to old archives. Now it turned out that the fix only worked, if there is still a temporary zip folder present that can be overwritten. If that file is not present, creating of a new zip archive failed.

This commit fixes this, by combining the CREATE and the OVERWRITE flag.