fix database query for getting a reheasal data
Avoid "select *" queries as they are not recommended. The explicit listing of selected fields fixes a bug where 3 database tables are JOINed and all of them contain a field "id". The resulting table has only one field "id" and it seems that its value (depending on the used SQL version and/or randomness) contains the "id" value of one of those three tables. If not the "id" field of the "rehearsal" table is used, then the returned table is incorrectly parsed and the rehearsal information on the start-page are missing some details (such as Ort/Conductor/Repertoire).